Rename AntiPattern to SmartPattern
